upstream: Share the proxy-chaining handshake state machine
The SOCKS5/HTTP CONNECT handshake this proxy speaks to an upstream
proxy was implemented twice, ~205 near-identical lines per loop that
differ only in the I/O primitive. Move the builders, the reply parse
switch, and the start logic into upstream.c behind a small step enum:
gwp_upstream_hs_start() and gwp_upstream_hs_on_reply() build requests
and parse replies without doing I/O, returning SEND / RECV / DONE (or
a negative errno); each loop arms the send/recv and runs its own
finish tail. gwp_upstream_dst_str() moves too.

The downstream fail-notify policy stays in each caller and is
unchanged: epoll still tells its client (upstream_fail) only on a
protocol-level rejection, signalled by the parser's *notify out-param;
io_uring passes NULL and never notifies. The only observable change is
that io_uring's upstream handshake-started and CONNECT-failed log
lines now carry the dst=%s field the epoll lines already had; no test
asserts on them.

Verified: full make test green on the io_uring build (both loops) and
the epoll-only build; upstream chaining (socks5/socks5h/http + auth
failure) clean under ASan/UBSan.
